+++ resources/download/logboost.ps1
@@ -1,26 +1,23 @@
-param (
-    [String] $ScriptRoot=$PSScriptRoot
-)
+. ".\resources\download\common.ps1"
 
-${ScriptRoot} = "${ScriptRoot}\resources\download"
-${ROOT_PATH} = Resolve-Path "${ScriptRoot}\..\..\"
-
-. "${ScriptRoot}\common.ps1"
+${ROOT_PATH} = "${PWD}"
 
 Write-DateLog "Start Sandbox to install or update TI for LogBoost." > "${ROOT_PATH}\log\logboost.txt"
 
-$mutex = New-Object System.Threading.Mutex($false, $mutexName)
-
-
 if (Test-Path -Path "${ROOT_PATH}\mount\Tools\logboost" ) {
-    (Get-Content "${ROOT_PATH}\resources\templates\generate_logboost.wsb.template").replace("__SANDBOX__", "${ROOT_PATH}") | Set-Content "${ROOT_PATH}\tmp\generate_logboost.wsb"
+    (Get-Content "${ROOT_PATH}\resources\templates\generate_logboost.wsb.template").replace("__SANDBOX__", "${ROOT_PATH}\") | Set-Content "${ROOT_PATH}\tmp\generate_logboost.wsb"
+    Start-Process "${ROOT_PATH}\tmp\generate_logboost.wsb"
 
-    $mutex.WaitOne() | Out-Null
-    & "${ROOT_PATH}\tmp\generate_logboost.wsb"
-    Start-Sleep 10
-    Remove-Item "${ROOT_PATH}\tmp\generate_logboost.wsb" | Out-Null
-
-    Stop-SandboxWhenDone "${ROOT_PATH}\mount\Tools\logboost\done" $mutex | Out-Null
+    do {
+        Start-Sleep 10
+        if (Test-Path -Path "${ROOT_PATH}\mount\Tools\logboost\done" ) {
+            Stop-Sandbox
+            Remove-Item "${ROOT_PATH}\tmp\generate_logboost.wsb" | Out-Null
+            Start-Sleep 1
+        }
+    } while (
+        tasklist | Select-String "(WindowsSandboxClient|WindowsSandboxRemote)"
+    )
 
     Write-DateLog "Logboost done." >> "${ROOT_PATH}\log\logboost.txt"
 } else {
