+++ resources/download/msys2.ps1
@@ -1,12 +1,8 @@
-param (
-    [String] $ScriptRoot=$PSScriptRoot
-)
+. ".\resources\download\common.ps1"
 
-$ScriptRoot = "$ScriptRoot\resources\download"
-${ROOT_PATH} = Resolve-Path "$ScriptRoot\..\..\"
+$ROOT_PATH = "${PWD}"
 
-. "${ScriptRoot}\common.ps1"
-
+Write-Output "${ROOT_PATH}"
 Write-DateLog "Setup MSYS2 and install packages in Sandbox." > ${ROOT_PATH}\log\msys2.txt
 
 if (! (Test-Path -Path "${ROOT_PATH}\tmp" )) {
@@ -29,20 +25,20 @@ if (! (Test-Path -Path "${ROOT_PATH}\mount\Tools\msys64" )) {
     New-Item -ItemType Directory -Force -Path "${ROOT_PATH}\mount\Tools\msys64" | Out-Null
 }
 
-# Create mutex and wait for it
-$mutex = New-Object System.Threading.Mutex($false, $mutexName)
-$mutex.WaitOne() | Out-Null
-
 # Create WSB file for MSYS2 and run it
-(Get-Content ${ROOT_PATH}\resources\templates\generate_msys2.wsb.template).replace('__SANDBOX__', "${ROOT_PATH}") | Set-Content "${ROOT_PATH}\tmp\generate_msys2.wsb"
-& "${ROOT_PATH}\tmp\generate_msys2.wsb"
+(Get-Content ${ROOT_PATH}\resources\templates\generate_msys2.wsb.template).replace('__SANDBOX__', "${ROOT_PATH}\") | Set-Content "${ROOT_PATH}\tmp\generate_msys2.wsb"
+Start-Process "${ROOT_PATH}\tmp\generate_msys2.wsb"
 
-# Wait for MSYS2 to start and then remove tmp WSB file
-Start-Sleep 10
-Remove-Item "${ROOT_PATH}\tmp\generate_msys2.wsb" | Out-Null
-
-# Wait for MSYS2 to finish
-Stop-SandboxWhenDone "${ROOT_PATH}\tmp\Tools\msys64\done" $mutex | Out-Null
+do {
+    Start-Sleep 10
+    if (Test-Path -Path "${ROOT_PATH}\tmp\Tools\msys64\done" ) {
+        Stop-Sandbox
+        Remove-Item "${ROOT_PATH}\tmp\generate_msys2.wsb" | Out-Null
+        Start-Sleep 1
+    }
+} while (
+    tasklist | Select-String "(WindowsSandboxClient|WindowsSandboxRemote)"
+)
 
 # Remove debug directory - use -recurse if there is any file created in the directory
 if (Test-Path -Path "${ROOT_PATH}\tmp\Tools\Debug" ) {
@@ -53,7 +49,7 @@ if (Test-Path -Path "${ROOT_PATH}\tmp\Tools\Debug" ) {
 Write-Output "C:/tmp/msys2 /tmp ntfs auto 0 0" >> "${ROOT_PATH}\tmp\Tools\msys64\etc\fstab"
 
 # Copy MSYS2 to Tools
-rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\Tools\msys64" "${ROOT_PATH}\mount\Tools\msys64"
+rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\Tools\msys64" "${ROOT_PATH}\mount\Tools\msys64" >> ${ROOT_PATH}\log\msys2.txt 2>&1
 Write-DateLog "MSYS2 and packages done." >> ${ROOT_PATH}\log\msys2.txt 2>&1
 
 Remove-Item -Recurse -Force "${ROOT_PATH}\tmp\Tools" 2>&1 | Out-Null
