+++ resources/download/node.ps1
@@ -1,14 +1,8 @@
-param (
-    [String] $ScriptRoot=$PSScriptRoot
-)
+. ".\resources\download\common.ps1"
 
-$ScriptRoot = "$ScriptRoot\resources\download"
-${ROOT_PATH} = Resolve-Path "$ScriptRoot\..\..\"
-
-. "${ScriptRoot}\common.ps1"
+$ROOT_PATH = "${PWD}"
 
 Write-DateLog "Setup Node and install npm packages in Sandbox." > ${ROOT_PATH}\log\npm.txt
-Write-DateLog "" > ${ROOT_PATH}\log\npm.txt
 
 $node_packages = @("box-js", "deobfuscator", "docsify-cli", "jsdom")
 
@@ -35,22 +29,25 @@ if (Test-Path -Path "${ROOT_PATH}\mount\Tools\node\node.txt" ) {
 }
 
 if ((Get-FileHash "${ROOT_PATH}\tmp\node\node.txt").Hash -ne (Get-FileHash ${CURRENT}).Hash) {
-    $mutex = New-Object System.Threading.Mutex($false, $mutexName)
-
     if (! (Test-Path -Path "${ROOT_PATH}\mount\Tools\node" )) {
         New-Item -ItemType Directory -Force -Path "${ROOT_PATH}\mount\Tools\node" | Out-Null
     }
 
-    (Get-Content ${ROOT_PATH}\resources\templates\generate_node.wsb.template).replace('__SANDBOX__', "${ROOT_PATH}") | Set-Content "${ROOT_PATH}\tmp\generate_node.wsb"
+    (Get-Content ${ROOT_PATH}\resources\templates\generate_node.wsb.template).replace('__SANDBOX__', "${ROOT_PATH}\") | Set-Content "${ROOT_PATH}\tmp\generate_node.wsb"
+    Start-Process "${ROOT_PATH}\tmp\generate_node.wsb"
+    
+    do {
+        Start-Sleep 10
+        if (Test-Path -Path "${ROOT_PATH}\tmp\node\done" ) {
+            Stop-Sandbox
+            Remove-Item "${ROOT_PATH}\tmp\generate_node.wsb" | Out-Null
+            Start-Sleep 1
+        }
+    } while (
+        tasklist | Select-String "(WindowsSandboxClient|WindowsSandboxRemote)"
+    )
 
-    $mutex.WaitOne() | Out-Null
-    & "${ROOT_PATH}\tmp\generate_node.wsb"
-    Start-Sleep 10
-    Remove-Item "${ROOT_PATH}\tmp\generate_node.wsb" | Out-Null
-
-    Stop-SandboxWhenDone "${ROOT_PATH}\tmp\node\done" $mutex | Out-Null
-
-    rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\node" "${ROOT_PATH}\mount\Tools\node"
+    rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\node" "${ROOT_PATH}\mount\Tools\node" >> ${ROOT_PATH}\log\npm.txt 2>&1
     Write-DateLog "Node and npm packages done." >> ${ROOT_PATH}\log\npm.txt 2>&1
 } else {
     Write-DateLog "Node and npm packages already installed and up to date." >> ${ROOT_PATH}\log\npm.txt 2>&1
