+++ resources/download/rust.ps1
@@ -1,15 +1,8 @@
-param (
-    [String] $ScriptRoot=$PSScriptRoot
-)
+. ".\resources\download\common.ps1"
 
-${ScriptRoot} = "${ScriptRoot}\resources\download"
-${ROOT_PATH} = Resolve-Path "${ScriptRoot}\..\..\"
-
-. "${ScriptRoot}\common.ps1"
+$ROOT_PATH = "${PWD}"
 
 Write-DateLog "Start Sandbox to install Rust based tools for dfirws." > "${ROOT_PATH}\log\rust.txt"
-
-$mutex = New-Object System.Threading.Mutex($false, $mutexName)
 
 # Requires gcc to compile
 ${CURRENT_VERSION_DFIR_TOOLKIT} = (curl --silent -L "https://crates.io/api/v1/crates/dfir-toolkit" | ConvertFrom-Json).crate.max_stable_version
@@ -49,16 +42,21 @@ if (! ${STATUS}) {
         Remove-Item "${ROOT_PATH}\tmp\cargo\done" | Out-Null
     }
 
-    (Get-Content "${ROOT_PATH}\resources\templates\generate_rust.wsb.template").replace("__SANDBOX__", "${ROOT_PATH}") | Set-Content "${ROOT_PATH}\tmp\generate_rust.wsb"
+    (Get-Content "${ROOT_PATH}\resources\templates\generate_rust.wsb.template").replace("__SANDBOX__", "${ROOT_PATH}\") | Set-Content "${ROOT_PATH}\tmp\generate_rust.wsb"
+    Start-Process "${ROOT_PATH}\tmp\generate_rust.wsb"
 
-    $mutex.WaitOne() | Out-Null
-    & "${ROOT_PATH}\tmp\generate_rust.wsb"
-    Start-Sleep 10
-    Remove-Item "${ROOT_PATH}\tmp\generate_rust.wsb" | Out-Null
+    do {
+        Start-Sleep 10
+        if (Test-Path -Path "${ROOT_PATH}\tmp\cargo\done" ) {
+            Stop-Sandbox
+            Remove-Item "${ROOT_PATH}\tmp\generate_rust.wsb" | Out-Null
+            Start-Sleep 1
+        }
+    } while (
+        tasklist | Select-String "(WindowsSandboxClient|WindowsSandboxRemote)"
+    )
 
-    Stop-SandboxWhenDone "${ROOT_PATH}\tmp\cargo\done" $mutex | Out-Null
-
-    rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\cargo" "${ROOT_PATH}\mount\Tools\cargo"
+    rclone.exe sync --verbose --checksum "${ROOT_PATH}\tmp\cargo" "${ROOT_PATH}\mount\Tools\cargo" >> "${ROOT_PATH}\log\rust.txt" 2>&1
     Remove-Item -Recurse -Force "${ROOT_PATH}\tmp\cargo" 2>&1 | Out-Null
 
     Write-DateLog "Rust tools done." >> "${ROOT_PATH}\log\rust.txt"
